+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [next]
 - BREAKING: Adding serialization with `serde` as a feature: `serde_feature`.
+- Adding `Debug` and `Serialize`/`Deserialize` traits to a couple of more types.
 
 ## [0.7.0] - 2020-07-02
 No changelog is kept up to and including this version.